@@ -2,12 +2,13 @@
   "name": "maxrects-packer",
   "version": "2.7.3",
   "description": "A max rectangle 2d bin packer for packing glyphs or images into multiple sprite-sheet/atlas",
+  "type": "module",
   "main": "dist/maxrects-packer.js",
   "module": "dist/maxrects-packer.mjs",
   "types": "dist/maxrects-packer.d.ts",
   "scripts": {
     "clean": "rimraf ./lib && rimraf ./dist",
-    "build": "tsc && rollup -c rollup.config.ts",
+    "build": "rollup -c",
     "build:clean": "npm run clean && npm run build",
     "doc": "typedoc && touch docs/.nojekyll",
     "doc:json": "typedoc --json docs/typedoc.json",
@@ -36,25 +37,33 @@
   },
   "homepage": "https://github.com/soimy/maxrects-packer#readme",
   "devDependencies": {
-    "@types/jest": "^27.4.0",
-    "@types/node": "^12.12.58",
+    "@rollup/plugin-terser": "^0.3.0",
+    "@types/jest": "^29.2.5",
+    "@types/node": "^18.11.18",
+    "@typescript-eslint/eslint-plugin": "^5.47.1",
+    "@typescript-eslint/eslint-plugin-tslint": "^5.48.0",
+    "@typescript-eslint/parser": "^5.47.1",
     "ascii-table": "0.0.9",
-    "coveralls": "^3.1.0",
-    "cz-conventional-changelog": "^2.1.0",
-    "gh-pages": "^2.2.0",
-    "jest": "^27.4.7",
-    "rimraf": "^2.7.1",
-    "rollup": "^1.32.1",
-    "rollup-plugin-ts": "^1.3.0",
-    "rollup-plugin-uglify-es": "0.0.1",
-    "standard-version": "^9.3.2",
-    "ts-jest": "^27.1.3",
-    "ts-node": "^8.10.2",
-    "tslib": "^1.13.0",
-    "tslint": "^5.20.1",
-    "tslint-config-standard": "^8.0.1",
-    "typedoc": "^0.22.11",
-    "typescript": "^3.9.7"
+    "coveralls": "^3.1.1",
+    "cz-conventional-changelog": "^3.3.0",
+    "esbuild": "^0.16.12",
+    "eslint": "^8.31.0",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsdoc": "^39.6.4",
+    "gh-pages": "^4.0.0",
+    "jest": "^29.3.1",
+    "rimraf": "^3.0.2",
+    "rollup": "^3.9.1",
+    "rollup-plugin-dts": "^5.1.0",
+    "rollup-plugin-esbuild": "^5.0.0",
+    "rollup-plugin-typescript2": "^0.34.1",
+    "standard-version": "^9.5.0",
+    "ts-jest": "^29.0.3",
+    "ts-node": "^10.9.1",
+    "tslib": "^2.4.1",
+    "tslint-config-standard": "^9.0.0",
+    "typedoc": "^0.23.23",
+    "typescript": "^4.9.4"
   },
   "config": {
     "commitizen": {
