@@ -1,6 +1,8 @@
-module.exports = {
+export default {
   preset: 'ts-jest',
-  testEnvironment: 'node',
+  // testEnvironment: 'node',
+  testEnvironment: 'jest-environment-node',
+  transform: {},
   verbose: true,
   coverageDirectory: './test/coverage'
 };
