@@ -11,7 +11,7 @@ const opt = {
     allowRotation: false,
     tag: false,
     exclusiveTag: true
-}
+};
 
 let packer;
 beforeEach(() => {
@@ -20,46 +20,46 @@ beforeEach(() => {
 
 describe("#add", () => {
     test("adds first element correctly", () => {
-        packer.add(1000, 1000, {number: 1});
-        expect(packer.bins[0].rects[0].data.number).toBe(1);
+        packer.add(1000, 1000, {num: 1});
+        expect(packer.bins[0].rects[0].data.num).toBe(1);
     });
 
     test("creates additional bin if element doesn't fit in existing bin", () => {
-        packer.add(1000, 1000, {number: 1});
-        packer.add(1000, 1000, {number: 2});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(1000, 1000, {num: 2});
         expect(packer.bins.length).toBe(2);
-        expect(packer.bins[1].rects[0].data.number).toBe(2);
+        expect(packer.bins[1].rects[0].data.num).toBe(2);
     });
 
     test("adds to existing bins if possible", () => {
-        packer.add(1000, 1000, {number: 1});
-        packer.add(1000, 1000, {number: 2});
-        packer.add(10, 10, {number: 3});
-        packer.add(10, 10, {number: 4});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(1000, 1000, {num: 2});
+        packer.add(10, 10, {num: 3});
+        packer.add(10, 10, {num: 4});
         expect(packer.bins.length).toBe(2);
     });
 
     test("adds to new bins after next() is called", () => {
-        packer.add(1000, 1000, {number: 1});
-        packer.add(1000, 1000, {number: 2});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(1000, 1000, {num: 2});
         packer.next();
-        packer.add(10, 10, {number: 3});
-        packer.add(10, 10, {number: 4});
+        packer.add(10, 10, {num: 3});
+        packer.add(10, 10, {num: 4});
         expect(packer.bins.length).toBe(3);
         expect(packer.bins[packer.bins.length - 1].rects.length).toBe(2);
     });
 
     test("adds to bins with tag matching on", () => {
         packer.options.tag = true;
-        packer.add(1000, 1000, {number: 1});
-        packer.add(10, 10, {number: 2});
-        packer.add(1000, 1000, {number: 3, tag: "one"});
-        packer.add(1000, 1000, {number: 4, tag: "one"});
-        packer.add(10, 10, {number: 5, tag: "one"});
-        packer.add(10, 10, {number: 6, tag: "one"});
-        packer.add(10, 10, {number: 7, tag: "two"});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(10, 10, {num: 2});
+        packer.add(1000, 1000, {num: 3, tag: "one"});
+        packer.add(1000, 1000, {num: 4, tag: "one"});
+        packer.add(10, 10, {num: 5, tag: "one"});
+        packer.add(10, 10, {num: 6, tag: "one"});
+        packer.add(10, 10, {num: 7, tag: "two"});
         packer.next();
-        packer.add(10, 10, {number: 8, tag: "two"});
+        packer.add(10, 10, {num: 8, tag: "two"});
         expect(packer.bins.length).toBe(5);
         expect(packer.bins[0].rects.length).toBe(2);
         expect(packer.bins[0].tag).toBeUndefined();
@@ -73,12 +73,12 @@ describe("#add", () => {
 
     test("adds to bins with tag matching disable", () => {
         packer.options.tag = false;
-        packer.add(1000, 1000, {number: 1});
-        packer.add(10, 10, {number: 2});
-        packer.add(1000, 1000, {number: 3, tag: "one"});
-        packer.add(10, 10, {number: 4, tag: "two"});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(10, 10, {num: 2});
+        packer.add(1000, 1000, {num: 3, tag: "one"});
+        packer.add(10, 10, {num: 4, tag: "two"});
         packer.next();
-        packer.add(10, 10, {number: 8, tag: "two"});
+        packer.add(10, 10, {num: 8, tag: "two"});
         expect(packer.bins.length).toBe(3);
         expect(packer.bins[0].tag).toBeUndefined();
         expect(packer.bins[1].tag).toBeUndefined();
@@ -120,8 +120,8 @@ describe("#add", () => {
     });
 
     test("allows oversized elements to be added", () => {
-        packer.add(1000, 1000, {number: 1});
-        packer.add(2000, 2000, {number: 2});
+        packer.add(1000, 1000, {num: 1});
+        packer.add(2000, 2000, {num: 2});
         expect(packer.bins.length).toBe(2);
         expect(packer.bins[1].rects[0].width).toBe(2000);
         expect(packer.bins[1].rects[0].oversized).toBe(true);
@@ -166,8 +166,8 @@ describe("#sort", () => {
 describe("#addArray", () => {
     test("adds multiple elements to bins", () => {
         let input = [
-            {width: 1000, height: 1000, data: {number: 1}},
-            {width: 1000, height: 1000, data: {number: 2}}
+            {width: 1000, height: 1000, data: {num: 1}},
+            {width: 1000, height: 1000, data: {num: 2}}
         ];
         packer.addArray(input);
         expect(packer.bins.length).toBe(2);
@@ -175,10 +175,10 @@ describe("#addArray", () => {
 
     test("adds the big rects first", () => {
         let input = [
-            {width: 600, height: 20, data: {number: 1}},
-            {width: 600, height: 20, data: {number: 2}},
-            {width: 1000, height: 1000, data: {number: 3}},
-            {width: 1000, height: 1000, data: {number: 4}}
+            {width: 600, height: 20, data: {num: 1}},
+            {width: 600, height: 20, data: {num: 2}},
+            {width: 1000, height: 1000, data: {num: 3}},
+            {width: 1000, height: 1000, data: {num: 4}}
         ];
         packer.addArray(input);
         expect(packer.bins.length).toBe(2);
@@ -186,10 +186,10 @@ describe("#addArray", () => {
 
     test("adds the big rects & big hash first", () => {
         let input = [
-            {width: 600, height: 20, data: {number: 1}, hash: "aaa"},
-            {width: 600, height: 20, data: {number: 2}, hash: "bbb"},
-            {width: 1000, height: 1000, data: {number: 3}, hash: "ccc"},
-            {width: 1000, height: 1000, data: {number: 4}, hash: "ddd"}
+            {width: 600, height: 20, data: {num: 1}, hash: "aaa"},
+            {width: 600, height: 20, data: {num: 2}, hash: "bbb"},
+            {width: 1000, height: 1000, data: {num: 3}, hash: "ccc"},
+            {width: 1000, height: 1000, data: {num: 4}, hash: "ddd"}
         ];
         packer.addArray(input);
         expect(packer.bins.length).toBe(2);
@@ -204,7 +204,7 @@ describe("#addArray", () => {
 
     test("add one element array", () => {
         let input = [
-            {width: 600, height: 20, data: {number: 1}, hash: "aaa"}
+            {width: 600, height: 20, data: {num: 1}, hash: "aaa"}
         ];
         packer.addArray(input); // test null array error
         expect(packer.bins.length).toBe(1);
@@ -215,10 +215,10 @@ describe("#save & load", () => {
     test("Load old bins and continue packing", () => {
         packer.options.tag = true;
         let input = [
-            {width: 512, height: 512, data: {number: 1}},
-            {width: 512, height: 512, data: {number: 2}, tag: "one"},
-            {width: 512, height: 512, data: {number: 3}, tag: "one"},
-            {width: 512, height: 512, data: {number: 4}},
+            {width: 512, height: 512, data: {num: 1}},
+            {width: 512, height: 512, data: {num: 2}, tag: "one"},
+            {width: 512, height: 512, data: {num: 3}, tag: "one"},
+            {width: 512, height: 512, data: {num: 4}},
         ];
         packer.addArray(input);
         expect(packer.bins.length).toBe(2);
@@ -239,17 +239,17 @@ describe("#save & load", () => {
 describe("misc functionalities", () => {
     test("passes padding through", () => {
         packer = new MaxRectsPacker(1024, 1024, 4, opt);
-        packer.add(500, 500, {number: 1});
-        packer.add(500, 500, {number: 1});
-        packer.add(500, 500, {number: 1});
+        packer.add(500, 500, {num: 1});
+        packer.add(500, 500, {num: 1});
+        packer.add(500, 500, {num: 1});
         expect(packer.bins[0].width).toBe(1004);
     });
 
     test("get bins dirty status", () => {
         packer = new MaxRectsPacker(1024, 1024, 4, opt);
-        packer.add(508, 508, {number: 1});
+        packer.add(508, 508, {num: 1});
         let tester1 = packer.add(new Rectangle(508, 508));
-        let tester2 = packer.add(508, 508, {number: 3});
+        let tester2 = packer.add(508, 508, {num: 3});
         expect(packer.dirty).toBe(true);
         for (let bin of packer.bins) bin.setDirty(false);
         expect(packer.dirty).toBe(false);
@@ -259,7 +259,7 @@ describe("misc functionalities", () => {
         tester2.height = 512;
         expect(packer.dirty).toBe(true);
     });
-    
+
     test("quick repack & deep repack", () => {
         packer = new MaxRectsPacker(1024, 1024, 0, {...opt, ...{ tag: true }});
         let rect = packer.add(1024, 512, {hash: "6"});
@@ -270,7 +270,7 @@ describe("misc functionalities", () => {
         packer.add({width: 512, height: 512, data: {hash: "1"}, tag: "one"});
         expect(packer.bins.length).toBe(3);
         for (let bin of packer.bins) bin.setDirty(false); // clean dirty status
-        rect.width = 512; // shrink width 
+        rect.width = 512; // shrink width
         packer.repack(); // quick repack
         expect(packer.bins.length).toBe(3);
         packer.repack(false); // deep repack
@@ -312,6 +312,6 @@ describe("misc functionalities", () => {
         // true overriding
         x = packer.add(398, 48, { allowRotation: true });
         expect(packer.bins.length).toBe(1);
-        expect(x.rot).toBe(true)
-    })
+        expect(x.rot).toBe(true);
+    });
 });
